fix: band recruit 합치기
