+++ src/main/java/org/example/buskmate/band/dto/band/BandListItemResponse.java
@@ -1,4 +1,4 @@
-package org.example.buskmate.band.dto;
+package org.example.buskmate.band.dto.band;
 
 import lombok.Builder;
 import lombok.Getter;
