+++ src/main/java/org/example/buskmate/band/dto/recruitpost/CreateRecruitPostRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.buskmate.recruit.post.dto;
+package org.example.buskmate.band.dto.recruitpost;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -7,11 +7,16 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class UpdateRecruitPostRequestDto {
+public class CreateRecruitPostRequestDto {
+    @NotBlank
+    private String bandId;
+
     @NotBlank
     @Size(max = 50)
     private String title;
+
     @NotBlank
     @Size(max = 2000)
     private String content;
+
 }
