+++ src/main/java/org/example/buskmate/band/dto/recruitpost/RecruitPostListDto.java
@@ -1,8 +1,8 @@
-package org.example.buskmate.recruit.post.dto;
+package org.example.buskmate.band.dto.recruitpost;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.buskmate.recruit.post.domain.RecruitPostStatus;
+import org.example.buskmate.band.domain.RecruitPostStatus;
 
 import java.time.LocalDateTime;
