+++ src/main/java/org/example/buskmate/band/dto/recruitpost/RecruitPostStatusResponseDto.java
@@ -1,9 +1,9 @@
-package org.example.buskmate.recruit.post.dto;
+package org.example.buskmate.band.dto.recruitpost;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.example.buskmate.recruit.post.domain.RecruitPostStatus;
+import org.example.buskmate.band.domain.RecruitPostStatus;
 
 @Getter
 @Builder
