+++ src/main/java/org/example/buskmate/band/dto/recruitpost/UpdateRecruitPostRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.buskmate.recruit.post.dto;
+package org.example.buskmate.band.dto.recruitpost;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -7,16 +7,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class CreateRecruitPostRequestDto {
-    @NotBlank
-    private String bandId;
-
+public class UpdateRecruitPostRequestDto {
     @NotBlank
     @Size(max = 50)
     private String title;
-
     @NotBlank
     @Size(max = 2000)
     private String content;
-
 }
