+++ src/main/java/org/example/buskmate/band/service/RecruitApplicationService.java
@@ -1,0 +1,4 @@
+package org.example.buskmate.band.service;
+
+public interface RecruitApplicationService {
+}
