+++ src/main/java/org/example/buskmate/band/service/RecruitApplicationServiceImpl.java
@@ -1,16 +1,16 @@
-package org.example.buskmate.recruit.application.service;
+package org.example.buskmate.band.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.buskmate.recruit.application.dto.RecruitApplyResponseDto;
-import org.example.buskmate.recruit.application.repository.RecruitApplicationRepository;
-import org.example.buskmate.recruit.post.repository.RecruitPostRepository;
+import org.example.buskmate.band.dto.recruitapplication.RecruitApplyResponseDto;
+import org.example.buskmate.band.repository.RecruitApplicationRepository;
+import org.example.buskmate.band.repository.RecruitPostRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 
-public class RecruitApplicationServiceImpl implements RecruitApplicationService{
+public class RecruitApplicationServiceImpl implements RecruitApplicationService {
     private final RecruitApplicationRepository recruitApplicationRepository;
     private final RecruitPostRepository recruitPostRepository;
