--- src/main/java/org/example/buskmate/recruit/application/service/RecruitApplicationService.java
@@ -1,4 +1,0 @@
-package org.example.buskmate.recruit.application.service;
-
-public interface RecruitApplicationService {
-}
